Move RoW PV function to locations directory
