+++ ocelot/transformations/__init__.py
@@ -7,10 +7,7 @@ from .activity_links import (
     manage_activity_links,
 )
 from .locations import relabel_global_to_row
-from .cleanup import (
-    drop_zero_pv_row_datasets,
-    ensure_all_datasets_have_production_volume,
-)
+from .cleanup import ensure_all_datasets_have_production_volume
 from .parameterization import (
     create_pv_parameters,
     fix_ecoinvent_parameters,
