+++ ocelot/transformations/cleanup.py
@@ -1,8 +1,5 @@
 # -*- coding: utf-8 -*-
-from ..data_helpers import (
-    production_volume,
-    reference_products_as_string,
-)
+from ..data_helpers import production_volume
 import logging
 import pprint
 
@@ -12,26 +9,6 @@ def ensure_all_datasets_have_production_volume(data):
     for ds in data:
         assert production_volume(ds), "Dataset does not have valid production volume:\n{}".format(pprint.pformat(ds))
     return data
-
-
-def drop_zero_pv_row_datasets(data):
-    """Drop datasets which have the location ``RoW`` and zero production volumes.
-
-    Zero production volumes occur when all inputs have been allocated to region-specific datasets."""
-    for ds in (x for x in data if x['location'] == 'RoW'):
-        if production_volume(ds) == 0:
-            logging.info({
-                'type': 'table element',
-                'data': (ds['name'], reference_products_as_string(ds))
-            })
-    return [ds for ds in data
-            if (ds['location'] != 'RoW' or production_volume(ds) != 0)]
-
-
-drop_zero_pv_row_datasets.__table__ = {
-    'title': 'Drop `RoW` datasets with zero production volumes',
-    'columns': ["Name", "Product(s)"]
-}
 
 
 def deparameterize(dataset):
